added view details classes and update indexUI
added the view details file that contains the view details classes for each event type.  Also added a function in the top_level_functions file to call the view details form as a top level pop up for a certian event.  Used the new functions in the top_level_functions file to add the view details function to the IndexUI class when event selected in tree view