added delete booking and UI for view total income
Added the methods needed in the DBAccess class and the update class to allow a user to delete a booking from the database.  Also added the basic UI for the View Total Income window.  Have also added functions on main menu button clicks to allow user to open the total income window.